♻️  Split connection::Reply into a separate file
